+++ tests/testthat/fixtures/coevfit_examples.R
@@ -19,7 +19,7 @@ warmup <- 50
 iter <- 50
 chains <- 1
 
-# fit model without distance matrix
+# fit model without longitude and latitude values
 coevfit_example_01 <-
   coev_fit(
     data = d,
@@ -39,9 +39,12 @@ coevfit_example_01 <-
     seed = 12345
   )
 
-# fit model with distance matrix
-dist_mat <- as.matrix(dist(rnorm(n)))
-rownames(dist_mat) <- colnames(dist_mat) <- d$id
+# fit model with longitude and latitude values
+lon_lat <- data.frame(
+  id = d$id,
+  longitude = runif(n, -180, 180),
+  latitude = runif(n, -90, 90)
+)
 coevfit_example_02 <-
   coev_fit(
     data = d,
@@ -51,7 +54,7 @@ coevfit_example_02 <-
     ),
     id = "id",
     tree = tree,
-    dist_mat = dist_mat,
+    lon_lat = lon_lat,
     chains = chains,
     iter_warmup = warmup,
     iter_sampling = iter,
@@ -236,6 +239,32 @@ coevfit_example_10 <-
     seed = 12345
   )
 
+# fit model with approximate gaussian processes
+d <- data.frame(
+  id = tree$tip.label,
+  x = rnorm(n),
+  y = rnorm(n),
+  longitude = runif(n, -180, 180),
+  latitude = runif(n, -90, 90)
+)
+coevfit_example_11 <-
+  coev_fit(
+    data = d,
+    variables = list(
+      x = "normal",
+      y = "normal"
+    ),
+    id = "id",
+    tree = tree,
+    lon_lat = lon_lat,
+    dist_k = 3,
+    chains = chains,
+    iter_warmup = warmup,
+    iter_sampling = iter,
+    adapt_delta = 0.99,
+    seed = 12345
+  )
+
 # update cmdstanr file locations
 update_file_location <- function(coevfit) {
   coevfit$fit$save_output_files(
@@ -256,6 +285,7 @@ suppressMessages({
   update_file_location(coevfit_example_08)
   update_file_location(coevfit_example_09)
   update_file_location(coevfit_example_10)
+  update_file_location(coevfit_example_11)
 })
 
 # save coevfit objects as rds files
@@ -279,3 +309,19 @@ save_coevfit_rds(coevfit_example_07)
 save_coevfit_rds(coevfit_example_08)
 save_coevfit_rds(coevfit_example_09)
 save_coevfit_rds(coevfit_example_10)
+save_coevfit_rds(coevfit_example_11)
+
+# cleanup
+rm(
+  coevfit_example_01,
+  coevfit_example_02,
+  coevfit_example_03,
+  coevfit_example_04,
+  coevfit_example_05,
+  coevfit_example_06,
+  coevfit_example_07,
+  coevfit_example_08,
+  coevfit_example_09,
+  coevfit_example_10,
+  coevfit_example_11
+)